backend/requirements.txt
django==2
django-debug-toolbar==1.8
gunicorn==19.4.5
whitenoise==3.3.1
djangorestframework==3.8.2
Pygments==2.2.0
PyMySQL==0.9.2
boto==2.48.0
django-pb-model==0.1.6
numpy==1.15.4
phonenumbers==8.10.4
psycopg2==2.7.7
requests==2.21.0
twilio==6.25.1
pyjwt==1.7.1
hyper==0.7.0
pathlib2==2.3.3
channels==2.2.0
channels-redis==2.4.0
Pillow==6.0.0
python-resize-image==1.1.18
APScheduler==3.6.1
channels==2.2.0
channels-redis==2.4.0
APScheduler==3.6.1
attrs==18.2.0
apns==2.0.1
apns2==0.5.0
cffi==1.11.5
Django==1.11.18
cryptography==2.4.2
django-rest==0.0.1
PySocks==1.6.8
urllib3==1.24.1
